Refactor: extract auth setup to a separate module
